feat: v3.8.29 — MCP 参数类型恢复 + 记忆 ID 时间戳 + 进入对话锁定最新消息
(1) MCP 图表工具 -32602 Invalid parameters 根治:
模型对深嵌套 schema (数组/对象/数字/布尔) 常生成 JSON 字符串字面量
(如 data 为 "[{...}]" 而非数组), 服务端类型校验失败。新增
restoreTypesBySchema: 按工具 inputSchema 声明的类型, 把字符串值
恢复为对应 JSON 原生类型 (array/object 解析回结构化, number/
integer 转数字, boolean 转布尔), 并向下递归 properties/items。
schema 缺失时保持原样不污染。解决所有 mcp__ 工具参数类型丢问题。

(2) 记忆 ID 改为创建时间戳:
AssistantMemory/MemoryEntity/MemoryDAO id 由 Int 自增改为 Long
时间戳 YYMMDDHHMMSS (如 260820213001), 同一秒新建冲突顺延+1。
记忆工具删除预览读 id 改 longOrNull (Int 溢出 12 位时间戳)。

(3) 进入对话默认锁定最新一条消息:
原 currentMessages.size+5 超界索引被 clamp 到最后一条顶部对齐,
最后一条为长文本(模型大段回复)时视口显示其开头而非真正末尾,
表现为锁定模型最后回复。改为 scrollToItem(lastIndex, scrollOffset=
Int.MAX_VALUE) 底部对齐, 锁定最新一条消息, 与发送者无关。

版本: versionCode 200 (锁定), versionName 3.8.29

+++ app/src/main/java/me/rerere/rikkahub/data/ai/tools/MemoryTools.kt
@@ -22,7 +22,7 @@ import me.rerere.rikkahub.data.model.AssistantMemory
 fun buildMemoryTools(
     json: Json,
     onCreation: suspend (String) -> AssistantMemory,
-    onUpdate: suspend (Int, String) -> AssistantMemory,
+    onUpdate: suspend (Long, String) -> AssistantMemory,
     onDelete: suspend (Int) -> Unit
 ): List<Tool> = listOf(
     Tool(
